task: #3589342 Refactor tests code via Rector, July 2026 - include deadCode prepared set
By: mondrake
By: dcam

+++ core/tests/Drupal/KernelTests/Core/Validation/UniqueValuesConstraintValidatorTest.php
@@ -54,7 +54,7 @@ class UniqueValuesConstraintValidatorTest extends KernelTestBase {
   public function testValidation(): void {
     // Create entity with two values for the testing field.
     $definition = [
-      'id' => (int) rand(0, getrandmax()),
+      'id' => rand(0, getrandmax()),
       'user_id' => 0,
       'field_test_text' => [
         'text1',
@@ -70,7 +70,7 @@ class UniqueValuesConstraintValidatorTest extends KernelTestBase {
 
     // Create another entity with two values for the testing field.
     $definition = [
-      'id' => (int) rand(0, getrandmax()),
+      'id' => rand(0, getrandmax()),
       'user_id' => 0,
       'field_test_text' => [
         'text3',
@@ -94,7 +94,7 @@ class UniqueValuesConstraintValidatorTest extends KernelTestBase {
 
     // Create another entity with two values, but one value is existing.
     $definition = [
-      'id' => (int) rand(0, getrandmax()),
+      'id' => rand(0, getrandmax()),
       'user_id' => 0,
       'field_test_text' => [
         'text5',
@@ -310,7 +310,7 @@ class UniqueValuesConstraintValidatorTest extends KernelTestBase {
   public function testValidationCaseInsensitive(): void {
     // Create entity with two values for the testing field.
     $definition = [
-      'id' => (int) rand(0, getrandmax()),
+      'id' => rand(0, getrandmax()),
       'user_id' => 0,
       'field_test_text' => [
         'text1',
@@ -324,7 +324,7 @@ class UniqueValuesConstraintValidatorTest extends KernelTestBase {
     // identical to other value, but with different capitalization which should
     // still trigger a validation error.
     $definition = [
-      'id' => (int) rand(0, getrandmax()),
+      'id' => rand(0, getrandmax()),
       'user_id' => 0,
       'field_test_text' => [
         'Text1',
